Introduce Integrations/Article_As_Main_Entity.php
